+++ src/main/java/onboarding/Problem2/Problem2.java
@@ -1,0 +1,9 @@
+package onboarding.Problem2;
+
+public class Problem2 {
+    public static String solution(String cryptogram) {
+        Decoder decodeQueue = new Decoder(cryptogram);
+        while(decodeQueue.removeDuplicatedCharacter());
+        return decodeQueue.toString();
+    }
+}
